refactor: rename context providers to use the ContextProvider suffix
