+++ src/components/Configure/ObjectManagementNav/index.tsx
@@ -5,8 +5,8 @@ import {
   Box, Divider, Tabs, Text,
 } from '@chakra-ui/react';
 
-import { useInstallIntegrationProps } from '../../../context/InstallIntegrationContext';
-import { useProject } from '../../../context/ProjectContext';
+import { useInstallIntegrationProps } from '../../../context/InstallIntegrationContextProvider';
+import { useProject } from '../../../context/ProjectContextProvider';
 import { capitalize } from '../../../utils';
 import { useObjectsConfigureState } from '../state/ConfigurationStateProvider';
 import { useHydratedRevision } from '../state/HydratedRevisionContext';
