+++ src/components/Configure/content/useMutateInstallation.tsx
@@ -1,8 +1,8 @@
-import { useApiKey } from '../../../context/ApiKeyProvider';
-import { useConnections } from '../../../context/ConnectionsContext';
+import { useApiKey } from '../../../context/ApiKeyContextProvider';
+import { useConnections } from '../../../context/ConnectionsContextProvider';
 import { useErrorState } from '../../../context/ErrorContextProvider';
-import { useInstallIntegrationProps } from '../../../context/InstallIntegrationContext';
-import { useProject } from '../../../context/ProjectContext';
+import { useInstallIntegrationProps } from '../../../context/InstallIntegrationContextProvider';
+import { useProject } from '../../../context/ProjectContextProvider';
 import { useSelectedObjectName } from '../ObjectManagementNav';
 import { useObjectsConfigureState } from '../state/ConfigurationStateProvider';
 import { useHydratedRevision } from '../state/HydratedRevisionContext';
