+++ src/components/Oauth/Salesforce/SalesforceOauthFlow.tsx
@@ -6,8 +6,8 @@
 import { useCallback, useState } from 'react';
 
 import { PROVIDER_SALESFORCE } from '../../../constants';
-import { useApiKey } from '../../../context/ApiKeyProvider';
-import { useProject } from '../../../context/ProjectContext';
+import { useApiKey } from '../../../context/ApiKeyContextProvider';
+import { useProject } from '../../../context/ProjectContextProvider';
 import OAuthPopup from '../../Connect/OAuthPopup';
 import { fetchOAuthCallbackURL } from '../fetchOAuthCallbackURL';
