+++ src/context/ProjectContextProvider.tsx
@@ -6,7 +6,7 @@ import { LoadingIcon } from '../assets/LoadingIcon';
 import { ErrorTextBox } from '../components/ErrorTextBox';
 import { api, Project } from '../services/api';
 
-import { useApiKey } from './ApiKeyProvider';
+import { useApiKey } from './ApiKeyContextProvider';
 import {
   ErrorBoundary, useErrorState,
 } from './ErrorContextProvider';
